refactor: TLY-65-app-api 모듈 테스틀 도메인 기반 패키지 구조로 개편
