+++ threadly-apps/app-api/src/test/java/com/threadly/post/controller/UpdatePostStatusApiTest.java
@@ -1,4 +1,4 @@
-package com.threadly.controller.post;
+package com.threadly.post.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
