+++ threadly-apps/app-api/src/test/java/com/threadly/post/controller/like/GetPostLikeApiTest.java
@@ -1,4 +1,4 @@
-package com.threadly.controller.post.like;
+package com.threadly.post.controller.like;
 
 import static com.threadly.utils.TestConstants.EMAIL_VERIFIED_USER_1;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,7 +6,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.threadly.CommonResponse;
 import com.threadly.ErrorCode;
-import com.threadly.controller.post.BasePostApiTest;
+import com.threadly.post.controller.BasePostApiTest;
 import com.threadly.post.engagement.GetPostEngagementApiResponse;
 import com.threadly.post.like.post.GetPostLikersApiResponse;
 import com.threadly.testsupport.fixture.posts.PostLikeFixtureLoader;
